+++ CHANGES.md
@@ -6,12 +6,14 @@ Grammars:
 - enh(swift) add support for `distributed` keyword [Marcus Ortiz][]
 - enh(xml) recognize Unicode letters instead of only ASCII letters in XML element and attribute names (#3256)[Martin Honnen][]
 - Added 3rd party Toit grammar to SUPPORTED_LANGUAGES [Serzhan Nasredin][]
+- Use substring() instead of deprecated substr() [Tobias Buschor][]
 - Added 3rd party Oak grammar to SUPPORTED_LANGUAGES [Tim Smith][]
 
 [Bradley Mackey]: https://github.com/bradleymackey
 [Marcus Ortiz]: https://github.com/mportiz08
 [Martin Honnen]: https://github.com/martin-honnen
 [Serzhan Nasredin]: https://github.com/snxx-lppxx
+[Tobias Buschor]: https://github.com/nuxodin/
 [Tim Smith]: https://github.com/timlabs
 
 ## Version 11.5.0